Add minItems property to feature_names (#15)
This PR adds minItems property to the feature_names arrays and sets it to 0. This makes it explicit that it is allowed for the array to be empty (e.g. in case of decision tree stumps). Furthermore, the update enforces the feature names to be distinct.

I also updated the examples to make sense, since we want only the feature names to be listed, which are really used by the trained models.